requirements.txt
# Core dependencies
streamlit>=1.28.0
openai-whisper>=20231117
librosa>=0.10.0
pydub>=0.25.1
soundfile>=0.12.1
torch>=2.0.0
numpy>=1.24.0

# Optional: for vocal isolation (uncomment to enable)
# spleeter>=2.4.0

# Audio format support
ffmpeg-python>=0.2.0
